add auth

+++ models/user.js
@@ -9,18 +9,16 @@ class User {
   // Static methods
 
   // Create a new user
-  static createUser(username, email) {
+  static createUser(email, password) {
     return db.query(
-      "INSERT INTO users (username, email, cart_id) VALUES (?, ?, ?)",
-      [username, email]
+      "INSERT INTO users (email, password) VALUES (?, ?)",
+      [email, password]
     );
   }
 
   // Find a user by ID
   static findById(userId) {
-    return db.query("SELECT * FROM users WHERE id = ?", [
-      userId,
-    ]);
+    return db.query("SELECT * FROM users WHERE id = ?", [userId]);
   }
 
   // Get all users
@@ -30,18 +28,19 @@ class User {
 
   // Delete a user by ID
   static deleteById(userId) {
-    return db.query("DELETE FROM users WHERE id = ?", [
-      userId,
-    ]);
+    return db.query("DELETE FROM users WHERE id = ?", [userId]);
   }
 
   // Update user email by ID
   static updateEmail(userId, newEmail) {
-    return db.query(
-      "UPDATE users SET email = ? WHERE id = ?",
-      [newEmail, userId]
-    );
+    return db.query("UPDATE users SET email = ? WHERE id = ?", [
+      newEmail,
+      userId,
+    ]);
+  }
+  static findByEmail(email) {
+    return db.execute("SELECT * FROM users WHERE email = ?", [email]);
+  }
 }
-};
 
 module.exports = User;
